requirements_cpu.txt
-f https://download.pytorch.org/whl/torch_stable.html torch==1.8.1+cpu torchvision==0.9.1+cpu torchaudio; sys_platform=="linux"
torch torchvision torchaudio; sys_platform == 'darwin'
torch torchvision torchaudio; sys_platform == 'win32'
librosa
tensorboard
pyyaml
pandas
tqdm
mutagen
torchlibrosa
optuna
addict